+++ policyengine_us/parameters/gov/states/ca/cdss/tanf/cash/resources/limit/vehicle.yaml
@@ -1,0 +1,14 @@
+description: California limits the CalWORKs to filers with vehicle value below this amount.
+
+values:
+  2023-01-01: 32_045
+
+metadata:
+  unit: currency-USD
+  period: year
+  label: California CalWORKs vehicle value limit
+  reference:
+  - title: Los Angeles Department DPSS CalWORKs
+    href: https://dpss.lacounty.gov/en/cash/calworks.html
+  - title: Los Angeles Department of Public Social Services E-Policy | Determining Value of Property – Vehicles
+    href: http://epolicy.dpss.lacounty.gov/epolicy/epolicy/server/general/projects_responsive/ePolicyMaster/index.htm?&area=general&type=responsivehelp&ctxid=&project=ePolicyMaster#t=mergedProjects/CalWORKs/CalWORKs/42-215_Determining_Value_of_Property_Vehicles/42-215_Determining_Value_of_Property_Vehicles.htm
