--- policyengine_us/parameters/gov/states/ca/cdss/tanf/child_care/eligibility/resource_limit/property/age_threshold.yaml
@@ -1,12 +1,0 @@
-description: California increases the CalWORKs Child Care property limit for households with at least one member of the household at or over this age threshold or disabled.
-
-values:
-  2023-01-01: 60
-
-metadata:
-  unit: year
-  period: year
-  label: California CalWORKs Child Care aged or disabled property value limit age threshold
-  reference:
-  - title: Los Angeles Department DPSS CalWORKs
-    href: https://dpss.lacounty.gov/en/cash/calworks.html
