--- policyengine_us/parameters/gov/states/ca/cdss/tanf/child_care/eligibility/resource_limit/property/base.yaml
@@ -1,12 +1,0 @@
-description: California limits CalWORKs Child Care to households with property value below this amount in a household with no aged or disabled members.
-
-values:
-  2023-01-01: 10_888
-
-metadata:
-  unit: currency-USD
-  period: year
-  label: California CalWORKs Child Care base property value limit
-  reference:
-  - title: Los Angeles Department DPSS CalWORKs
-    href: https://dpss.lacounty.gov/en/cash/calworks.html
