--- policyengine_us/parameters/gov/states/ca/cdss/tanf/child_care/eligibility/resource_limit/vehicle.yaml
@@ -1,12 +1,0 @@
-description: California limits the CalWORKs Child Care to filers with vehicle value below this amount.
-
-values:
-  2023-01-01: 32_045
-
-metadata:
-  unit: currency-USD
-  period: year
-  label: California CalWORKs Child Care vehicle value limit
-  reference:
-  - title: Los Angeles Department DPSS CalWORKs
-    href: https://dpss.lacounty.gov/en/cash/calworks.html
