+++ policyengine_us/tests/policy/baseline/gov/states/ca/cdss/tanf/cash/eligibility/ca_tanf_vehicle_value_eligible.yaml
@@ -4,7 +4,7 @@
     state_code: CA
     household_vehicles_value: 32_045
   output:
-    ca_calworks_child_care_vehicle_value_eligible: true
+    ca_tanf_vehicle_value_eligible: true
 
 - name: Vehicle value ineligible
   period: 2023
@@ -12,4 +12,4 @@
     state_code: CA
     household_vehicles_value: 32_046
   output:
-    ca_calworks_child_care_vehicle_value_eligible: false
+    ca_tanf_vehicle_value_eligible: false
