--- policyengine_us/tests/policy/baseline/gov/states/ca/cdss/tanf/child_care/eligibility/ca_calworks_child_care_property_value_eligible.yaml
@@ -1,99 +1,0 @@
-- name: No aged or disabled members, eligible 
-  period: 2023
-  input:
-    people:
-      person1:
-        age: 59
-        is_disabled: false
-      person2:
-        age: 59
-        is_disabled: false
-    spm_units:
-      spm_unit:
-        members: [person1, person2]
-        spm_unit_cash_assets: 10_888
-    households:
-      household:
-        members: [person1, person2]
-  output:
-    ca_calworks_child_care_property_value_eligible: true
-
-- name: No aged or disabled members, ineligible 
-  period: 2023
-  input:
-    people:
-      person1:
-        age: 59
-        is_disabled: false
-      person2:
-        age: 59
-        is_disabled: false
-    spm_units:
-      spm_unit:
-        members: [person1, person2]
-        spm_unit_cash_assets: 10_889
-    households:
-      household:
-        members: [person1, person2]
-  output:
-    ca_calworks_child_care_property_value_eligible: false
-
-- name: One aged member, eligible
-  period: 2023
-  input:
-    people:
-      person1:
-        age: 60
-        is_disabled: false
-      person2:
-        age: 59
-        is_disabled: false
-    spm_units:
-      spm_unit:
-        members: [person1, person2]
-        spm_unit_cash_assets: 16_333
-    households:
-      household:
-        members: [person1, person2]
-  output:
-    ca_calworks_child_care_property_value_eligible: true
-
-- name: One disabled member, eligible
-  period: 2023
-  input:
-    people:
-      person1:
-        age: 59
-        is_disabled: false
-      person2:
-        age: 59
-        is_disabled: true
-    spm_units:
-      spm_unit:
-        members: [person1, person2]
-        spm_unit_cash_assets: 16_333
-    households:
-      household:
-        members: [person1, person2]
-  output:
-    ca_calworks_child_care_property_value_eligible: true
-
-- name: Aged and disabled member, ineligible
-  period: 2023
-  input:
-    people:
-      person1:
-        age: 60
-        is_disabled: false
-      person2:
-        age: 59
-        is_disabled: true
-    spm_units:
-      spm_unit:
-        members: [person1, person2]
-        spm_unit_cash_assets: 16_334
-    households:
-      household:
-        members: [person1, person2]
-  output:
-    ca_calworks_child_care_property_value_eligible: false
